fix: don't prefer local ports from other addresses when dialing
This address may already be in-use (on that other address) somewhere
else.

Thanks to @schomatis for figuring this out.

fixes #1611
