+++ p2p/net/reuseport/transport.go
@@ -31,5 +31,5 @@ type Transport struct {
 type network struct {
 	mu        sync.RWMutex
 	listeners map[*listener]struct{}
-	dialer    dialer
+	dialer    *dialer
 }
